refactor: rename usesession to usecurrentsession
